InstaSellUtil utils, RestrictSell (now InstaSellRestrictions) rewrite +
ADDED: InstaSellUtil, InstaSellItem (might need to just use OrderData instead)
CHANGED: RestrictSell -> InstaSellRestrictions, RestrictSellControl -> SellRestrictionControl, unfinished InstaSellRestrictions rewrite, now InstaSellHighlight uses InstaSellUtil class instead of doing it locally, findItemPrice now uses optionals
